pyproject.toml
[project]
name = "bioactive-ml"
version = "0.1.0"
description = "Meal Shield — Multi-task GNN + Topological Data Analysis for food peptide bioactivity prediction"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Miyu Horiuchi" }]

dependencies = [
    "torch>=2.0",
    "torch-geometric>=2.4",
    "rdkit-pypi>=2023.3",
    "ripser>=0.6",
    "persim>=0.3",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "networkx>=3.0",
    "requests>=2.31",
    "fair-esm>=2.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.0",
]
scraping = [
    "gdown>=5.0",
]
dev = [
    "pytest>=7.0",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["*.py"]
